Change DataType for PCIeRxBytes and PCIeTxBytes
'''
Changes required as data type in PDI corresponding to this is changed from uint32_t to uint64_t

'''

Fixes nvbug https://nvbugspro.nvidia.com/bug/5171760

+++ redfish-core/lib/processor.hpp
@@ -2106,7 +2106,7 @@ inline void getProcessorPerformanceData(
                 (property.first == "PCIeRXBytes"))
             {
                 auto propName = property.first;
-                const uint32_t* val = std::get_if<uint32_t>(&property.second);
+                const uint64_t* val = std::get_if<uint64_t>(&property.second);
                 if (val == nullptr)
                 {
                     BMCWEB_LOG_DEBUG("Get  pcie bytes property failed");
